Adjust transfer structure
